Legg til behandlingstype: Tilbakekreving under "lag ny" Ikke vis årsak, eller krav-mottatt dersom tilbakekreving er valgt. Hvis vi lager tilbakekrevingsbehandling - lukk modal og informer via toast på behandlingsoversikt siden. Reload tilbakekrevingsbehandlinger og vis fram i behandlingsoversikt.
