+++ src/frontend/App/typer/behandlingstype.ts
@@ -3,6 +3,7 @@ export enum Behandlingstype {
     REVURDERING = 'REVURDERING',
     BLANKETT = 'BLANKETT',
     TEKNISK_OPPHØR = 'TEKNISK_OPPHØR',
+    TILBAKEKREVING = 'TILBAKEKREVING',
 }
 
 export const behandlingstypeTilTekst: Record<Behandlingstype, string> = {
@@ -10,4 +11,5 @@ export const behandlingstypeTilTekst: Record<Behandlingstype, string> = {
     REVURDERING: 'Revurdering',
     BLANKETT: 'Blankettbehandling',
     TEKNISK_OPPHØR: 'Teknisk opphør',
+    TILBAKEKREVING: 'Tilbakekreving',
 };
