+++ src/frontend/App/typer/toast.ts
@@ -1,9 +1,11 @@
 export enum EToast {
     VEDTAK_UNDERKJENT = 'VEDTAK_UNDERKJENT',
     BEHANDLING_HENLAGT = 'BEHANDLING_HENLAGT',
+    TILBAKEKREVING_OPPRETTET = 'TILBAKEKREVING_OPPRETTET',
 }
 
 export const toastTilTekst: Record<EToast, string> = {
     VEDTAK_UNDERKJENT: 'Vedtak underkjent',
     BEHANDLING_HENLAGT: 'Behandlingen er henlagt',
+    TILBAKEKREVING_OPPRETTET: 'Tilbakekrevingsbehandling opprettet',
 };
