+++ src/frontend/Komponenter/Personoversikt/Behandlingsoversikt.tsx
@@ -23,14 +23,13 @@ import { useApp } from '../../App/context/AppContext';
 import DataViewer from '../../Felles/DataViewer/DataViewer';
 import { PartialRecord } from '../../App/typer/common';
 import { ToggleName } from '../../App/context/toggles';
-import { AlertStripeFeil } from 'nav-frontend-alertstriper';
+import Alertstripe, { AlertStripeFeil } from 'nav-frontend-alertstriper';
 import { useToggles } from '../../App/context/TogglesContext';
 import { Knapp } from 'nav-frontend-knapper';
 import { Systemtittel } from 'nav-frontend-typografi';
 import { BehandlingStatus } from '../../App/typer/behandlingstatus';
 import { EtikettInfo } from 'nav-frontend-etiketter';
-import RevurderingModal from './RevurderingModal';
-import Alertstripe from 'nav-frontend-alertstriper';
+import LagBehandlingModal from './LagBehandlingModal';
 import {
     TilbakekrevingBehandling,
     TilbakekrevingBehandlingsresultatstype,
@@ -73,7 +72,9 @@ const Behandlingsoversikt: React.FC<{ fagsakId: string }> = ({ fagsakId }) => {
     const [fagsak, settFagsak] = useState<Ressurs<Fagsak>>(byggTomRessurs());
     const [tekniskOpphørFeilet, settTekniskOpphørFeilet] = useState<boolean>(false);
     const [kanStarteRevurdering, settKanStarteRevurdering] = useState<boolean>(false);
-    const [visRevurderingvalg, settVisRevurderingvalg] = useState<boolean>(false);
+    const [visLagBehandlingModal, settVisLagBehandlingModal] = useState<boolean>(false);
+    const [hentTilbakekrevingsbehandlinger, settHentTilbakekrevingsbehandlinger] =
+        useState<boolean>(true);
     const [feilFagsakHenting, settFeilFagsakHenting] = useState<string>();
     const [tilbakekrevingBehandlinger, settTilbakekrevingbehandlinger] = useState<
         Ressurs<TilbakekrevingBehandling[]>
@@ -113,12 +114,13 @@ const Behandlingsoversikt: React.FC<{ fagsakId: string }> = ({ fagsakId }) => {
     };
 
     useEffect(() => {
-        if (fagsakId) {
+        if (fagsakId && hentTilbakekrevingsbehandlinger) {
             hentFagsak();
             hentTilbakekrevingBehandlinger();
+            settHentTilbakekrevingsbehandlinger(false);
         }
         // eslint-disable-next-line
-    }, [fagsakId]);
+    }, [fagsakId, hentTilbakekrevingsbehandlinger]);
 
     useEffect(() => {
         if (fagsak.status === RessursStatus.SUKSESS) {
@@ -156,14 +158,15 @@ const Behandlingsoversikt: React.FC<{ fagsakId: string }> = ({ fagsakId }) => {
                         tilbakekrevingBehandlinger={tilbakekrevingBehandlinger}
                     />
                     {kanStarteRevurdering && (
-                        <KnappMedMargin onClick={() => settVisRevurderingvalg(true)}>
+                        <KnappMedMargin onClick={() => settVisLagBehandlingModal(true)}>
                             Opprett ny behandling
                         </KnappMedMargin>
                     )}
-                    <RevurderingModal
-                        visModal={visRevurderingvalg}
-                        settVisModal={settVisRevurderingvalg}
+                    <LagBehandlingModal
+                        visModal={visLagBehandlingModal}
+                        settVisModal={settVisLagBehandlingModal}
                         fagsakId={fagsakId}
+                        settHentTilbakekrevinger={settHentTilbakekrevingsbehandlinger}
                     />
                     {toggles[ToggleName.TEKNISK_OPPHØR] && (
                         <KnappMedMargin onClick={() => gjørTekniskOpphør()}>
